+++ Editor/PropertyDrawers/TextStyleDrawer.cs
@@ -28,6 +28,7 @@ namespace XCharts
 #else
                 PropertyField(prop, "m_Font");
 #endif
+                PropertyField(prop, "m_Wrap");
                 PropertyField(prop, "m_Rotate");
                 PropertyField(prop, "m_Offset");
                 PropertyField(prop, "m_Color");